Refactor, bugfix, print methods

+++ MatrixC.cpp
@@ -9,9 +9,9 @@ MatrixC::~MatrixC()
 {
 }
 
-void MatrixC::calculateMatrixC(Jacobian jacobian, double c, double ro)
+void MatrixC::calculateMatrixC(Jacobian jacobian, long double c, long double ro)
 {
-    double tempMatrix[4][4][4];
+    long double tempMatrix[4][4][4];
 
     for(int i = 0; i < 4; i++)
     {
@@ -23,8 +23,6 @@ void MatrixC::calculateMatrixC(Jacobian jacobian, double c, double ro)
             tempMatrix[i][j][3] = jacobian.N[3][j] * jacobian.N[3][i] * jacobian.detJacobian[3] * c * ro;
 
             C[i][j] = tempMatrix[i][j][0] + tempMatrix[i][j][1] + tempMatrix[i][j][2] + tempMatrix[i][j][3];
-
-            // std::cout << C[i][j] << std::endl;
         }
     }
 }
