+++ MatrixC.h
@@ -4,10 +4,10 @@
 class MatrixC
 {
 public:
-    double C[4][4], dNdx[4][4], dNdy[4][4], dNdxT[4][4][4], dNdyT[4][4][4], dNdxTdetJ[4][4][4], dNdyTdetJ[4][4][4], sum[4][4][4];
+    long double C[4][4], dNdx[4][4], dNdy[4][4], dNdxT[4][4][4], dNdyT[4][4][4], dNdxTdetJ[4][4][4], dNdyTdetJ[4][4][4], sum[4][4][4];
 
     MatrixC();
     ~MatrixC();
 
-    void calculateMatrixC(Jacobian jacobian, double c, double ro);
+    void calculateMatrixC(Jacobian jacobian, long double c, long double ro);
 };
